Add quiet option to suppress banner
This option suppresses the banner output, which is quite verbose and annoying
when running on many plugins at once (for example, in a test suite).  It would
be more conventional to not spam the console like this at all, but instead,
this commit adds a suppression option to preserve previous behavior.